requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
pymongo

# Authentication & Security
passlib[bcrypt]
python-jose[cryptography]

# Environment & Configuration
python-dotenv

# AI & ML Services
sentence-transformers
azure-ai-documentintelligence
google-generativeai

# Document Processing
python-docx==1.1.0
pdfplumber
docxtpl

# HTTP & API
httpx
requests
aiofiles

# Data Processing
pydantic==2.7.1
json5
rapidfuzz

# File Upload
python-multipart

# Template Engine
jinja2

# Storage
azure-storage-blob